# Conflicts:
#	README.md
#	src/engine/localevent.cpp
#	src/engine/system.h
#	src/fheroes2/game/game_highscores.cpp
#	src/fheroes2/system/settings.cpp

+++ src/fheroes2/agg/bin_info.h
@@ -22,8 +22,6 @@
 #define H2BIN_FRM_H
 
 #include "settings.h"
-
-#include <map>
 #include <vector>
 
 namespace Bin_Info
@@ -76,7 +74,7 @@ namespace Bin_Info
         int32_t troopCountOffsetLeft;
         int32_t troopCountOffsetRight;
         std::vector<Point> projectileOffset;
-        std::vector<double> projectileAngles;
+        std::vector<float> projectileAngles;
         std::vector<float> idlePriority;
         std::vector<uint32_t> unusedIdleDelays;
         uint32_t idleAnimationCount;
